Local modifications

+++ verl/workers/rollout/vllm_rollout/vllm_rollout_spmd.py
@@ -146,12 +146,14 @@ class vLLMRollout(BaseRollout):
         if config.get("limit_images", None):  # support for multi-image data
             engine_kwargs["limit_mm_per_prompt"] = {"image": config.get("limit_images")}
 
+        #assert 1 == 2, config.dtype
         self.inference_engine = LLM(
             model=model_path,
             enable_sleep_mode=True,
             tensor_parallel_size=tensor_parallel_size,
             distributed_executor_backend="external_launcher",
-            dtype=config.dtype,
+            #dtype=config.dtype,
+            dtype="bfloat16",
             enforce_eager=config.enforce_eager,
             gpu_memory_utilization=config.gpu_memory_utilization,
             disable_custom_all_reduce=True,
